worker/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Google Cloud
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-cloud-aiplatform==1.42.1

# Anthropic Claude
anthropic==0.40.0

# Image Processing
Pillow==10.2.0
rembg==2.0.56
numpy==1.26.3
opencv-python-headless==4.9.0.80

# Utilities
python-dotenv==1.0.0
